+++ device/options.go
@@ -73,14 +73,11 @@ type Options struct {
 	Now func() time.Time
 
 	// WRPSourceCheck defines behavior around checking the Source field in WRP messages originating
-	// from this device. The current check cases are:
-	// 1) Source is empty: leaving this field empty suggests very little chance of underlying bad intent
-	//    so the field is populated with the device's canonical ID (i.e. mac:112233445566).
-	// 2) Canonical ID can't be parsed from Source: a bad actor could be behind this behavior and thus the
-	//    message is dropped on the floor when the check type is "enforce".
-	// 3) Canonical ID doesn't match that of the established websocket connection: this is the most obvious case in
-	//    which a bad actor could have a device pretend sending messages oh behalf of another.
-	// Note: when the check type is "monitor", no messages are dropped but they are logged as part of the "wrp_source_check"
+	// from devices. All the following are cases of an invalid WRP wrt the source:
+	// 1) Source is empty.
+	// 2) Canonical ID can't be parsed from Source.
+	// 3) Canonical ID doesn't match that of the established websocket connection.
+	// Note: when the check type is "monitor", no messages are dropped but they are logged as an error and update the "wrp_source_check"
 	// counter.
 	WRPSourceCheck wrpSourceCheckConfig
 }