+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added eventKey to service discovery metric [#509](https://github.com/xmidt-org/webpa-common/pull/509)
 
 ## [v1.10.6]
 ### Fixed