+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Fixed
-- Fixed serviceEndpoints unit tests failing [#510](https://github.com/xmidt-org/webpa-common/pull/510)
+- Fixed serviceEndpoints unit tests failing. [#510](https://github.com/xmidt-org/webpa-common/pull/510)
+
+### Changed
+- Add configurable check for the source of inbound (device => cloud service) WRP messages. [#507] (https://github.com/xmidt-org/webpa-common/pull/507)
 
 ## [v1.10.6]
 ### Fixed