+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Add configurable check for the source of inbound (device => cloud service) WRP messages. [#507] (https://github.com/xmidt-org/webpa-common/pull/507)
 
+## Changed 
+- Populate empty inbound WRP.content_type field with application/octet-stream [#508](https://github.com/xmidt-org/webpa-common/pull/508)
+
 ## [v1.10.6]
 ### Fixed
 - Fixed ServiceEndpoints not faning out when a datacenter is empty [#504](https://github.com/xmidt-org/webpa-common/pull/504)