+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.10.7]
 ### Added
 - Added eventKey to service discovery metric. [#509](https://github.com/xmidt-org/webpa-common/pull/509)
 - Added `trust` label to hardware_model gauge. [#512](https://github.com/xmidt-org/webpa-common/pull/512)
@@ -144,7 +146,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - The first official release. We will be better about documenting changes 
    moving forward.
 
-[Unreleased]: https://github.com/xmidt-org/webpa-common/compare/v1.10.6...HEAD
+[Unreleased]: https://github.com/xmidt-org/webpa-common/compare/v1.10.7...HEAD
+[v1.10.7]: https://github.com/xmidt-org/webpa-common/compare/v1.10.6...v1.10.7
 [v1.10.6]: https://github.com/xmidt-org/webpa-common/compare/v1.10.5...v1.10.6
 [v1.10.5]: https://github.com/xmidt-org/webpa-common/compare/v1.10.4...v1.10.5
 [v1.10.4]: https://github.com/xmidt-org/webpa-common/compare/v1.10.3...v1.10.4